requirements.txt
python-telegram-bot[ext]>=20.0,<22.0
google-generativeai>=0.8.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
dateparser>=1.2.0
python-dotenv>=1.0.0
thefuzz[speedup]>=0.22.0
